% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gihyo_document.R
\name{gihyo_document}
\alias{gihyo_document}
\title{Gihyo Flavored Markdown}
\usage{
gihyo_document(md_extensions = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{md_extensions}{Markdown extensions to be added or removed from the
default definition or R Markdown. See the \code{\link[rmarkdown]{rmarkdown_format}} for
additional details.}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{Arguments passed to \code{md_document}}
}
\description{
Gihyo Flavored Markdown
}
