% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-gfm.R
\name{style_gfm}
\alias{style_gfm}
\title{Style R code blocks in gihyo-flavored markdown file}
\usage{
style_gfm(path, ...)
}
\arguments{
\item{path}{Paths to markdown files}

\item{...}{Arguments passed to \code{styler::style_file()}}
}
\description{
Style R code blocks in gihyo-flavored markdown file
}
